requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
opencv-python
django
djangorestframework
